Replace old xml profiler with new one
